requirements/base.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-decouple>=3.8
redis>=4.5.0
celery>=5.3.0
coverage>=7.0.0
factory-boy>=3.3.0
freezegun>=1.2.0
django-celery-beat>=2.7.0
gunicorn>=21.0.0
whitenoise>=6.0.0
APScheduler==3.10.4
django-apscheduler==0.6.2